FIT: support 64 bit addresses + sha256 hash

+++ NyaFs/ImageFormat/Types/ImageInfo.cs
@@ -13,8 +13,8 @@ namespace NyaFs.ImageFormat.Types
 
         public string Name = "";
 
-        public uint DataLoadAddress = 0;
-        public uint EntryPointAddress = 0;
+        public UInt64 DataLoadAddress = 0;
+        public UInt64 EntryPointAddress = 0;
 
         public ImageInfo Clone()
         {
